server/requirements.txt
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
flask-sqlalchemy>=3.0.0
mysql-connector-python>=8.0.0